ml_pipeline/requirements.txt
# ML Pipeline Dependencies
# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.42.1
google-cloud-functions==1.16.0
google-cloud-logging==3.8.0

# Machine Learning
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Audio Processing
soundfile>=0.12.1
resampy>=0.4.2
pydub>=0.25.1

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0